Move mock server related codes to a class
